Verify whether manual bounds input is used and correct
The check identifies that the input vector is numeric
and of length 4. An exception stops the function.
A successful check results in setting the boolean
TRUE for the new object 'use_manual_bounds'. This
new object will later nullify the use of the default
choice for the bounds, even if they exist and
'use.calmet.bounds' was set to TRUE, not set to
FALSE, or otherwise not set at all by the user.